+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # emqtt-bench changelog
 
+## 0.4.31
+* New `--ssl-version` to enforce TLS version and implies ssl is enabled.
+* QoE logging now logs TCP handshake latency during TLS handshake ( emqtt 1.14.0).
+* QoE logging now logs each publish msg' end to end latency if `--payload-hdrs=ts` is set by both subscriber and publisher. 
+* Dump TLS secrets per connecion to SSLKEYLOGFILE specifed by envvar SSLKEYLOGFILE for TLS traffic decryption.  (TLS and QUIC)
+* Now build release for arm64.
+* Now build release for el7 with newer build image.
+
 ## 0.4.30
 
 * Enhanced QoE trackings, add supports for commands `conn` and `pub`.